% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_util.R
\name{geom_lbl}
\alias{geom_lbl}
\title{geom_lbl}
\usage{
geom_lbl(..., family = theme_get()$text$family, size = 3, label.size = 0)
}
\arguments{
\item{...}{Additional parameters passed to \code{geom_label}}

\item{family}{Font family. Defaults to theme-defined family.}

\item{size}{Font size. Defaults to 3.}

\item{label.size}{Label size. Defaults to 0.}
}
\description{
Helper for \code{geom_label} with some defaults.
}
