% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_util.R
\name{geom_txt}
\alias{geom_txt}
\title{geom_txt}
\usage{
geom_txt(..., family = theme_get()$text$family, size = 3,
  colour = "#2b2b2b")
}
\arguments{
\item{...}{Passed to \code{geom_text}.}

\item{family}{Font family.  Defaults to theme-defined family.}

\item{size}{Font size. Defaults to 3.}

\item{colour}{Font colour. Defaults to \code{#2b2b2b}}
}
\description{
Helper for \code{geom_text} with some defaults.
}
