% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{h_colors}
\alias{h_colors}
\title{h_colors}
\format{An object of class \code{list} of length 14.}
\usage{
h_colors
}
\description{
List of colors for Harvard color palette.
}
\keyword{datasets}
