% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicify_names.R
\name{nicify_names}
\alias{nicify_names}
\title{Nicify names}
\usage{
nicify_names(df, style = c("snake_case"), abbrev = FALSE)
}
\arguments{
\item{df}{Dataframe}

\item{style}{Naming style convention.  Currently only 'snake_case' is supported.}

\item{abbrev}{Logical. Whether to use common abbreviations such as 'pct' for '\%'}
}
\description{
Nicify names
}
