% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_h}
\alias{theme_h}
\title{H ggplot theme}
\usage{
theme_h(..., flip = FALSE)
}
\arguments{
\item{...}{Arguments passed to \code{theme_minimal}}

\item{flip}{If \code{TRUE}, flips the coordinates so the grid lines go
vertically.}
}
\description{
Complete H theme.
}
