% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_h_}
\alias{theme_h_}
\title{Partial H theme}
\usage{
theme_h_(..., flip = FALSE)
}
\description{
Slightly modified version of \code{theme_nyt_}. Adds additional spacing under
title and subtitle elements; title is not bold and is colored
}

