% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_nyt}
\alias{theme_nyt}
\title{NYT ggplot theme}
\usage{
theme_nyt(..., flip = FALSE)
}
\arguments{
\item{...}{Parameters passed to \code{theme_minimal}}

\item{flip}{If \code{TRUE}, flips the coordinates so the grid lines go
vertically.}
}
\description{
Complete theme based on \code{theme_minimal} that applies NYT-style. Font to
be used should be specified as \code{base_family}.
}
\details{
Original post at \url{http://rud.is/b/2016/03/16/supreme-annotations/}
}

