% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{theme_nyt_}
\alias{theme_nyt_}
\title{Partial NYT theme}
\usage{
theme_nyt_(..., flip = FALSE)
}
\arguments{
\item{...}{Unused.}

\item{flip}{Whether to flip the axes}
}
\description{
Partial NYT theme
}

