% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{theme_nyt}
\alias{theme_nyt}
\title{NYT ggplot theme}
\usage{
theme_nyt(..., flip = FALSE)
}
\arguments{
\item{...}{Parameters passed to \code{theme_minimal}}

\item{flip}{Whether to flip the axes}
}
\description{
NYT ggplot theme
}

